Refactor : 공통 예외 PREFIX 설정
