--- src/main/java/lotto/Lotto.java
@@ -1,30 +1,0 @@
-package lotto;
-
-import camp.nextstep.edu.missionutils.Randoms;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Set;
-import java.util.TreeSet;
-
-public class Lotto implements LottoMachine {
-    private final Set<Integer> numbers;
-
-    public Lotto(List<Integer> numbers) {
-        validate(numbers);
-        this.numbers = new TreeSet<>();
-    }
-
-    private void validate(List<Integer> numbers) {
-        if (numbers.size() != 6) {
-            throw new IllegalArgumentException("[ERROR] 로또 번호는 6개여야 합니다.");
-        }
-    }
-
-    @Override
-    public void run() {
-        List<Integer> results = Randoms.pickUniqueNumbersInRange(1, 45, 6);
-        numbers.addAll(results);
-    }
-
-    // TODO: 추가 기능 구현
-}
