+++ src/main/java/lotto/view/OutputView.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.view;
 
 public class OutputView {
